content-pipeline/requirements.txt
feedparser>=6.0
anthropic>=0.39.0
python-dotenv>=1.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
